improved kmeans sample speed

+++ samples/fr/dynamo/samples/kmeans/Centroid.java
@@ -2,21 +2,21 @@ package fr.dynamo.samples.kmeans;
 
 public class Centroid {
 
-  private float sumX;
-  private float sumY;
+  private double sumX;
+  private double sumY;
   private int count;
 
-  public void add(float x, float y){
+  public void add(double x, double y){
     count++;
     sumX += x;
     sumY += y;
   }
 
-  public float get_x(){
+  public double get_x(){
     return sumX / count;
   }
 
-  public float get_y(){
+  public double get_y(){
     return sumY / count;
   }
 }
